different runtime for unit testing & benchmarking
