+++ pallets/bootstrap/Cargo.toml
@@ -22,12 +22,14 @@ sp-std = { default-features = false, version = '4.0.0-dev' , git = "https://gith
 sp-runtime = { git = "https://github.com/mangata-finance/substrate", default-features = false, branch = "mangata-dev" }
 sp-arithmetic= { git = "https://github.com/mangata-finance/substrate", default-features = false, branch = "mangata-dev" }
 frame-benchmarking = { default-features = false, version = '4.0.0-dev' , git = "https://github.com/mangata-finance/substrate", branch = "mangata-dev" }
+sp-bootstrap = { path = "../../primitives/bootstrap", default-features=false }
 
 [dev-dependencies]
 sp-core = { default-features = false, version = '4.1.0-dev' , git = "https://github.com/mangata-finance/substrate", branch = "mangata-dev" }
 sp-io = { default-features = false, version = '4.0.0-dev' , git = "https://github.com/mangata-finance/substrate", branch = "mangata-dev" }
 orml-traits = { default-features = false, version="0.4.1-dev" , git = "https://github.com/mangata-finance/open-runtime-module-library", branch = "mangata-dev" }
-pallet-xyk = { path = "../xyk", default-features = false}
+pallet-xyk = { path = "../xyk" }
+sp-bootstrap = { path = "../../primitives/bootstrap", default-features=false }
 pallet-issuance = { path = "../issuance", default-features = false}
 pallet-assets-info = { path = "../assets-info", default-features = false}
 mockall = "0.11.0"
@@ -45,5 +47,4 @@ std = [
     'frame-system/std',
     'orml-tokens/std',
 	"frame-benchmarking/std",
-	"pallet-xyk/std",
 ]
