+++ pallets/xyk/Cargo.toml
@@ -26,7 +26,7 @@ pallet-issuance = { default-features = false, path = "../issuance" }
 pallet-bootstrap = { default-features = false, path = "../bootstrap" }
 sp-arithmetic = { default-features = false, version = '4.0.0-dev' , git = "https://github.com/mangata-finance/substrate", branch = "mangata-dev" }
 libm = { default-features = false, git = "https://github.com/rust-lang/libm", rev="2f3fc968f43d345f9b449938d050a9ea46a04c83"}
-
+sp-bootstrap = { path = "../../primitives/bootstrap", default-features=false }
 orml-tokens = {  default-features = false, version = '0.4.1-dev' , git = "https://github.com/mangata-finance/open-runtime-module-library", branch = "mangata-dev" }
 
 pallet-assets-info = { path = '../assets-info', default-features = false, version = '2.0.0' }
