+++ runtime/Cargo.toml
@@ -165,6 +165,7 @@ std = [
 	"pallet-treasury/std",
 	"pallet-assets-info/std",
 	"pallet-xyk/std",
+	"pallet-bootstrap/std",
     "xyk-runtime-api/std",
 
 	"artemis-core/std",
